requirements.txt
numpy
tensorflow-gpu
ray
jupyter
requests
pillow
psutil
setproctitle
torch
torchvision
cython
scipy
pandas
mmcv